Move some functions to Composer
